Integrating Datepicker with other components
