+++ app/components/master-filter.js
@@ -2,8 +2,6 @@ import Component from '@ember/component';
 import { task, timeout } from 'ember-concurrency';
 import config from 'travis/config/environment';
 import { inject as service } from '@ember/service';
-
-
 export default Component.extend({
   api: service(),
   store: service(),
