+++ app/routes/travis-ci-insights.js
@@ -1,4 +1,4 @@
 import TravisRoute from 'travis/routes/basic';
 export default TravisRoute.extend({
-  needsAuth: true,
+  needsAuth: false,
 });
