+++ app/templates/components/aggregate-matrices.hbs
@@ -1,4 +1,4 @@
-<div class="container" data-test-aggregate-matrices-container repo={{ selectedReposAg}}>
+<div class="container" data-test-aggregate-matrices-container>
    <div class="flex-container aggregate-value">
         <div class="builds col-md-2" data-test-builds>
             <div data-test-builds-label>Builds</div>
