+++ app/templates/components/master-filter.hbs
@@ -1,4 +1,3 @@
-
 <div class="wrapper-all-filters" data-test-filter-container>
   {{#basic-dropdown as |dd|}}
     {{#dd.trigger class="repo-filter-label"}}
@@ -33,4 +32,3 @@
     {{/dd.content}}
   {{/basic-dropdown}}
 </div>
-{{yield selectedRepos}}
